all integer typedefs moved to ggnfs.h; ssize_t removed; #if should not be used inside assembler .s files
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@62 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/lasieve4/piii/basemath.c
@@ -17,24 +17,24 @@
 #include <sys/types.h>
 #include "lasieve-asm.h"
 
-extern unsigned long montgomery_inv_n;
-extern unsigned long *montgomery_modulo_n;
-extern unsigned long montgomery_modulo_R2[NMAX_ULONGS];
-extern unsigned long montgomery_modulo_R4[NMAX_ULONGS];
-extern unsigned long montgomery_ulongs;
+extern uint32_t montgomery_inv_n;
+extern uint32_t *montgomery_modulo_n;
+extern uint32_t montgomery_modulo_R2[NMAX_ULONGS];
+extern uint32_t montgomery_modulo_R4[NMAX_ULONGS];
+extern uint32_t montgomery_ulongs;
 
-extern void (*asm_mulmod) (unsigned long *, unsigned long *, unsigned long *);
-extern void (*asm_add2) (unsigned long *, unsigned long *);
-extern void (*asm_diff) (unsigned long *, unsigned long *, unsigned long *);
-extern void (*asm_zero) (unsigned long *);
-extern void (*asm_copy) (unsigned long *, unsigned long *);
-extern void (*asm_add2_ui) (unsigned long *, unsigned long);
-extern void (*asm_sub) (unsigned long *, unsigned long *, unsigned long *);
-extern void (*asm_sub_n) (unsigned long *, unsigned long *);
-extern void (*asm_half) (unsigned long *);
+extern void (*asm_mulmod) (uint32_t *, uint32_t *, uint32_t *);
+extern void (*asm_add2) (uint32_t *, uint32_t *);
+extern void (*asm_diff) (uint32_t *, uint32_t *, uint32_t *);
+extern void (*asm_zero) (uint32_t *);
+extern void (*asm_copy) (uint32_t *, uint32_t *);
+extern void (*asm_add2_ui) (uint32_t *, uint32_t);
+extern void (*asm_sub) (uint32_t *, uint32_t *, uint32_t *);
+extern void (*asm_sub_n) (uint32_t *, uint32_t *);
+extern void (*asm_half) (uint32_t *);
 
 /**************************************************/
-int asm_cmp64(unsigned long * a, unsigned long * b)
+int asm_cmp64(uint32_t * a, uint32_t * b)
 /**************************************************/
 {
   if (a[0] != b[0])
@@ -45,7 +45,7 @@ int asm_cmp64(unsigned long * a, unsigned long * b)
 }
 
 /**************************************************/
-int asm_cmp(unsigned long * a, unsigned long * b)
+int asm_cmp(uint32_t * a, uint32_t * b)
 /**************************************************/
 { long i;
 
@@ -56,9 +56,9 @@ int asm_cmp(unsigned long * a, unsigned long * b)
 }
 
 /**************************************************/
-void gcd64(unsigned long * gcd, unsigned long * a, unsigned long * b)
+void gcd64(uint32_t * gcd, uint32_t * a, uint32_t * b)
 /**************************************************/
-{ unsigned long r[2], bb[2], aa[2];
+{ uint32_t r[2], bb[2], aa[2];
 
   if (!(a[0] | a[1])) {
     gcd[0] = b[0];
@@ -103,9 +103,9 @@ void gcd64(unsigned long * gcd, unsigned long * a, unsigned long * b)
 }
 
 /**************************************************
-void gcd(unsigned long * gcd, unsigned long * a, unsigned long * b)
+void gcd(uint32_t * gcd, uint32_t * a, uint32_t * b)
 **************************************************
-{ unsigned long r[NMAX_ULONGS], bb[NMAX_ULONGS], aa[NMAX_ULONGS];
+{ uint32_t r[NMAX_ULONGS], bb[NMAX_ULONGS], aa[NMAX_ULONGS];
   long i;
 
   for (i = 0; i < montgomery_ulongs; i++)
@@ -159,9 +159,9 @@ void gcd(unsigned long * gcd, unsigned long * a, unsigned long * b)
 }*/
 
 /**************************************************/
-void asm_half_old(unsigned long * a)
+void asm_half_old(uint32_t * a)
 /**************************************************/
-{ unsigned long c, n_half[NMAX_ULONGS];
+{ uint32_t c, n_half[NMAX_ULONGS];
   long i;
 
   for (i = 0; i < montgomery_ulongs - 1; i++)
@@ -180,12 +180,12 @@ void asm_half_old(unsigned long * a)
 }
 
 /**************************************************/
-int asm_invert(unsigned long * res, unsigned long * b)
+int asm_invert(uint32_t * res, uint32_t * b)
 /**************************************************/ 
 { long i, f1, len;
-  unsigned long t1[NMAX_ULONGS], t2[NMAX_ULONGS];
-  unsigned long v1[NMAX_ULONGS], v2[NMAX_ULONGS];
-  unsigned long n_half[NMAX_ULONGS];
+  uint32_t t1[NMAX_ULONGS], t2[NMAX_ULONGS];
+  uint32_t v1[NMAX_ULONGS], v2[NMAX_ULONGS];
+  uint32_t n_half[NMAX_ULONGS];
 
   for (i = 0; i < montgomery_ulongs; i++)
     if (b[i])
